organizing code for constructing metadata
